tdd-guide/pointer-tdd/p.go
package p

import (
	"errors"
	"fmt"
)

type Bitcoin int

type Wallet struct {
	balance Bitcoin
}

var InsufficientFundsError = errors.New("cannot withdraw, insufficient funds")

func (w *Wallet) Deposit(num Bitcoin) {
	w.balance += num
}

func (w Wallet) Balance() Bitcoin {
	return w.balance
}

func (b Bitcoin) String() string {
	return fmt.Sprintf("%d BTC", b)
}

func (w *Wallet) Withdraw(num Bitcoin) error {
	if w.balance < num {
		return InsufficientFundsError
	}
	w.balance -= num
	return nil
}